fix firestore repository
