+++ pages/_app.tsx
@@ -2,6 +2,7 @@ import type {AppProps} from "next/app"
 import {ThemeProvider} from "styled-components"
 import {GlobalStyle} from "styles/GlobalStyle";
 import {themes} from "styles/theme";
+import "utils/firebase";
 
 function MyApp({Component, pageProps}: AppProps) {
   return (
